areas_put : passing code for name collision bug

+++ lambdas/public_api/areas/areas_put.py
@@ -42,7 +42,15 @@ def update_area(event, context):
                         new_trail_ids = [trail_id for trail_id in other_trail_ids if trail_id not in trail_ids]
                         area_table.put_item(Item={"name": other_area_name, "trail_ids": new_trail_ids})
 
-        if new_name:
+        if new_name and new_name != original_name:
+            existing = area_table.query(
+                KeyConditionExpression=Key("name").eq(new_name),
+                Limit=1
+            )["Items"]
+
+            if existing:
+                raise ValueError(f"Already found area with name [{new_name}]")
+
             area_table.delete_item(Key={"name": original_name})
         else:
             new_name = original_name
